move throttle_error to Handler and use time.time() instead of datetime.now()
